+++ interfaces/lib/vpc/interfaces.d.ts
@@ -1,8 +1,11 @@
-export interface VPCConfig {
+import { StackProps } from "aws-cdk-lib";
+import { VpcProps } from "aws-cdk-lib/aws-ec2/lib/vpc";
+export interface VPCConfig extends StackProps, VpcProps {
     stackName: string;
     stackDescription: string;
     natGateways: number;
-    cidrMask: number;
+    cidrRange?: string;
+    maxAzs: number;
     kubernetesClustersToTag: string[];
     vpcIdSSM: string;
     vpcIdSSMDescription: string;