+++ interfaces/lib/vpc/interfaces.ts
@@ -1,8 +1,12 @@
-export interface VPCConfig {
+import { StackProps } from "aws-cdk-lib";
+import { VpcProps } from "aws-cdk-lib/aws-ec2/lib/vpc";
+
+export interface VPCConfig extends StackProps, VpcProps  {
     stackName: string;
     stackDescription: string;
     natGateways: number;
-    cidrMask: number;
+    cidrRange?: string;
+    maxAzs: number;
     kubernetesClustersToTag: string[];
     vpcIdSSM: string;
     vpcIdSSMDescription: string;
@@ -12,4 +16,4 @@ export interface VPCConfig {
     publicSubnetSSMDescription: string
     isolatedSubnetSSM: string;
     isolatedSubnetSSMDescription: string;
-}
+  }