+++ lib/vpc/vpc.ts
@@ -1,4 +1,4 @@
-import { Stack, StackProps, Tags } from 'aws-cdk-lib';
+import { Stack, Tags } from 'aws-cdk-lib';
 import { SubnetType, Vpc } from 'aws-cdk-lib/aws-ec2';
 import { StringListParameter, StringParameter } from 'aws-cdk-lib/aws-ssm';
 import { Construct } from 'constructs';
@@ -8,10 +8,10 @@ export class VPCStack extends Stack {
   vpc: Vpc;
   config: VPCConfig;
 
-  constructor(scope: Construct, id: string, config: VPCConfig, props?: StackProps) {
+  constructor(scope: Construct, id: string, props: VPCConfig) {
     super(scope, id, props);
+    this.config = props;
 
-    this.config = config;
     this.createVPC();
     this.addTags();
     this.createParams();
@@ -19,8 +19,8 @@ export class VPCStack extends Stack {
 
   createVPC(): void {
     this.vpc = new Vpc(this, 'VPC', {
-      maxAzs: 4,
-      cidr: '10.0.0.0/16',
+      maxAzs: this.config.maxAzs || 4,
+      cidr: this.config.cidrRange || '10.0.0.0/16',
       natGateways: this.config.natGateways,
       subnetConfiguration: [
         {
@@ -29,7 +29,7 @@ export class VPCStack extends Stack {
         },
         {
           name: 'Private',
-          subnetType: SubnetType.PRIVATE_WITH_NAT,
+          subnetType: SubnetType.PRIVATE_WITH_EGRESS,
         },
         {
           name: 'Protected',
@@ -44,24 +44,22 @@ export class VPCStack extends Stack {
   }
 
   createParams(): void {
+
     const vpcIdParameter = new StringParameter(this, 'VPCId', {
       stringValue: this.vpc.vpcId,
       parameterName: this.config.vpcIdSSM,
       description: this.config.vpcIdSSMDescription,
     });
-
     const privateSubnetParameter = new StringListParameter(this, 'PrivateSubnetIds', {
       stringListValue: this.vpc.privateSubnets.map((subnet) => subnet.subnetId),
       parameterName: this.config.privateSubnetSSM,
       description: this.config.privateSubnetSSMDescription,
     });
-
     const publicSubnetParameter = new StringListParameter(this, 'PublicSubnetIds', {
       stringListValue: this.vpc.publicSubnets.map((subnet) => subnet.subnetId),
       parameterName: this.config.publicSubnetSSM,
       description: this.config.publicSubnetSSMDescription,
     });
-
     const isolatedSubnetSSM = new StringListParameter(this, 'IsolatedSubnetIds', {
       stringListValue: this.vpc.isolatedSubnets.map((subnet) => subnet.subnetId),
       parameterName: this.config.isolatedSubnetSSM,
@@ -73,16 +71,14 @@ export class VPCStack extends Stack {
     this.vpc.privateSubnets.forEach((subnet) => {
       Tags.of(subnet).add('kubernetes.io/role/internal-elb', '1');
     });
-
     this.vpc.publicSubnets.forEach((subnet) => {
       Tags.of(subnet).add('kubernetes.io/role/elb', '1');
     });
-
-    this.vpc.privateSubnets.forEach((subnet) => {
-      const ar: string[] = JSON.parse(this.config.kubernetesClustersToTag as unknown as string);
+    Object.entries(this.vpc.privateSubnets).forEach(([idx, subnet]) => {
+      const ar: string[] = this.config.kubernetesClustersToTag || []
       ar.forEach((cluster) => {
         Tags.of(subnet).add(`kubernetes.io/cluster/${cluster}`, 'shared');
       });
-    });
+    })
   }
 }